+++ lib/frontend/wp/users/refreshAuthToken.js
