+++ lib/frontend/wp/users/registerUser.js
